refactor: create myHistory package and add get watch history use case in it
